Interaction with the API in a separate file. Updating takes place via Websockets. The graph is rounded
According to the SRP, the implementation of API data acquisition is hidden in a separate file. In the API itself there is a limit on 1 open Websocket, which makes it impossible to launch 2 clients at the same time, it is desirable to fix this. Because of the Websockets, information comes almost instantly. The columns of the graph were rounded and expanded. The graph rendering call has been changed.